+++ src/running/command/runbms.py
@@ -23,6 +23,7 @@ minheap_multiplier: float
 remote_host: Optional[str]
 skip_oom: Optional[int]
 skip_timeout: Optional[int]
+skip_log_compression: bool = False
 plugins: Dict[str, Any]
 resume: Optional[str]
 
@@ -42,6 +43,8 @@ def setup_parser(subparsers):
     f.add_argument("--skip-timeout", type=int)
     f.add_argument("--resume", type=str)
     f.add_argument("--workdir", type=Path)
+    f.add_argument("--skip-log-compression", action="store_true",
+                   help="Skip compressing log files")
 
 
 def getid() -> str:
@@ -116,7 +119,7 @@ def run_benchmark_with_config(c: str, b: Benchmark, runbms_dir: Path, size: Opti
     mod_b = b.attach_modifiers(mods)
     mod_b = b.attach_modifiers(b.get_runtime_specific_modifiers(runtime))
     if size is not None:
-        mod_b = mod_b.attach_modifiers([runtime.get_heapsize_modifier(size)])
+        mod_b = mod_b.attach_modifiers(runtime.get_heapsize_modifiers(size))
     if fd:
         prologue = get_log_prologue(runtime, mod_b)
         fd.write(prologue.encode("ascii"))
@@ -133,8 +136,10 @@ def run_benchmark_with_config(c: str, b: Benchmark, runbms_dir: Path, size: Opti
 
 
 def get_filename_no_ext(bm: Benchmark, hfac: Optional[float], size: Optional[int], config: str) -> str:
+    # If we have / in benchmark names, replace with -.
+    safe_bm_name = bm.name.replace("/", "-")
     return "{}.{}.{}.{}.{}".format(
-        bm.name,
+        safe_bm_name,
         # plotty uses "^(\w+)\.(\d+)\.(\d+)\.([a-zA-Z0-9_\-\.\:\,]+)\.log\.gz$"
         # to match filenames
         hfac_str(hfac) if hfac is not None else "0",
@@ -288,11 +293,14 @@ def run_one_benchmark(
         p.end_benchmark(hfac, size, bm)
     for j, c in enumerate(configs):
         log_filename = get_filename(bm, hfac, size, c)
+        # Check that this is not a dry-run and we have actually executed this
+        # config for a particular benchmark/hfac (method parameters)
         if not is_dry_run() and ever_ran[j]:
-            subprocess.check_call([
-                "gzip",
-                log_dir / log_filename
-            ])
+            if not skip_log_compression:
+                subprocess.check_call([
+                    "gzip",
+                    log_dir / log_filename
+                ])
     print()
 
 
@@ -362,6 +370,8 @@ def run(args):
         skip_oom = args.get("skip_oom")
         global skip_timeout
         skip_timeout = args.get("skip_timeout")
+        global skip_log_compression
+        skip_log_compression = args.get("skip_log_compression")
         # Load from configuration file
         global configuration
         configuration = Configuration.from_file(