+++ src/running/runtime.py
@@ -1,5 +1,5 @@
-from running.modifier import JVMArg, Modifier, JSArg
-from typing import Any, Dict, Union
+from running.modifier import JVMArg, Modifier, JSArg, EnvVar
+from typing import Any, Dict, List, Union
 from pathlib import Path
 import logging
 from running.util import register
@@ -19,7 +19,7 @@ class Runtime(object):
     def get_executable(self) -> Union[str, Path]:
         raise NotImplementedError
 
-    def get_heapsize_modifier(self, size: int) -> Modifier:
+    def get_heapsize_modifiers(self, size: int) -> List[Modifier]:
         raise NotImplementedError
 
     def is_oom(self, _output: bytes) -> bool:
@@ -60,13 +60,13 @@ class JVM(Runtime):
     def __str__(self):
         return "JVM {}".format(self.name)
 
-    def get_heapsize_modifier(self, size: int) -> Modifier:
+    def get_heapsize_modifiers(self, size: int) -> List[Modifier]:
         size_str = "{}M".format(size)
         heapsize = JVMArg(
             name="heap{}".format(size_str),
             val="-Xms{} -Xmx{}".format(size_str, size_str)
         )
-        return heapsize
+        return [heapsize]
 
     def is_oom(self, output: bytes) -> bool:
         for pattern in [b"Allocation Failed", b"OutOfMemoryError", b"ran out of memory", b"panicked at 'Out of memory!'"]:
@@ -141,14 +141,14 @@ class D8(JavaScriptRuntime):
     def __str__(self):
         return "{} d8 {}".format(super().__str__(), self.executable)
 
-    def get_heapsize_modifier(self, size: int) -> Modifier:
+    def get_heapsize_modifiers(self, size: int) -> List[Modifier]:
         size_str = "{}".format(size)
         heapsize = JSArg(
             name="heap{}".format(size_str),
             val="--initial-heap-size={} --max-heap-size={}".format(
                 size_str, size_str)
         )
-        return heapsize
+        return [heapsize]
 
     def is_oom(self, output: bytes) -> bool:
         # The format is "Fatal javascript OOM in ..."
@@ -162,14 +162,14 @@ class SpiderMonkey(JavaScriptRuntime):
     def __str__(self):
         return "{} SpiderMonkey {}".format(super().__str__(), self.executable)
 
-    def get_heapsize_modifier(self, size: int) -> Modifier:
+    def get_heapsize_modifiers(self, size: int) -> List[Modifier]:
         size_str = "{}".format(size)
         # FIXME doesn't seem to be working
         heapsize = JSArg(
             name="heap{}".format(size_str),
             val="--available-memory={}".format(size_str)
         )
-        return heapsize
+        return [heapsize]
 
     def is_oom(self, output: bytes) -> bool:
         # FIXME not sure how to check for OOM for SpiderMonkey yet
@@ -181,15 +181,62 @@ class JavaScriptCore(JavaScriptRuntime):
     def __str__(self):
         return "{} JavaScriptCore {}".format(super().__str__(), self.executable)
 
-    def get_heapsize_modifier(self, size: int) -> Modifier:
+    def get_heapsize_modifiers(self, size: int) -> List[Modifier]:
         size_str = "{}".format(size)
         # FIXME doesn't seem to be working
         heapsize = JSArg(
             name="heap{}".format(size_str),
             val="--gcMaxHeapSize={}".format(size_str)
         )
-        return heapsize
+        return [heapsize]
 
     def is_oom(self, output: bytes) -> bool:
         # FIXME not sure how to check for OOM for JavaScriptCore yet
+        return False
+
+
+class Julia(Runtime):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.executable: Path
+        self.executable = Path(kwargs["executable"])
+        if not self.executable.exists():
+            logging.warning(
+                "Julia executable {} doesn't exist".format(self.executable))
+        self.executable = self.executable.absolute()
+
+    def get_executable(self) -> Path:
+        return self.executable
+
+    def __str__(self):
+        return "Julia {} {}".format(self.name, self.executable)
+
+
+@register(Runtime)
+class JuliaMMTK(Julia):
+    def get_heapsize_modifiers(self, size: int) -> List[Modifier]:
+        # size in MB
+        size_str = "{}".format(size)
+        min = EnvVar(name="minheap{}".format(size_str),
+                     var="MMTK_MIN_HSIZE", val=size_str)
+        max = EnvVar(name="maxheap{}".format(size_str),
+                     var="MMTK_MAX_HSIZE", val=size_str)
+        return [min, max]
+
+    def __str__(self):
+        return "{} with MMTk".format(super().__str__())
+
+    def is_oom(self, output: bytes) -> bool:
+        return b"Out of Memory!" in output
+
+
+@register(Runtime)
+class JuliaStock(Julia):
+    def get_heapsize_modifiers(self, size: int) -> List[Modifier]:
+        return []
+
+    def __str__(self):
+        return "{} stock version".format(super().__str__())
+
+    def is_oom(self, output: bytes) -> bool:
         return False