+++ docs/src/changelog.md
@@ -1,6 +1,18 @@
 # Changelog
 ## Unreleased
 ### Added
+#### Modifiers
+- `JuliaArg`
+
+#### Runtimes
+- `JuliaMMTK`
+- `JuliaStock`
+
+#### Benchmark Suites
+- `JuliaGCBenchmarks`
+
+#### Commands
+- `runbms` gains an extra argument, `--skip-log-compression`, to skip compressing log files with `gzip`.
 
 ### Changed
 #### Base Configurations